COMMON: Move Language and Platform functionality into separate files
