+++ backends/platform/dc/dc.h
@@ -29,6 +29,8 @@
 #include "backends/audiocd/default/default-audiocd.h"
 #include "backends/fs/fs-factory.h"
 #include "audio/mixer_intern.h"
+#include "common/language.h"
+#include "common/platform.h"
 #ifdef DYNAMIC_MODULES
 #include "backends/plugins/dynamic-plugin.h"
 #endif
