+++ common/util.cpp
@@ -110,174 +110,6 @@ bool parseBool(const String &val, bool &valAsBool) {
 #pragma mark -
 
 
-const LanguageDescription g_languages[] = {
-	{ "zh-cn", "zh_CN", "Chinese (China)", ZH_CNA },
-	{    "zh", "zh_TW", "Chinese (Taiwan)", ZH_TWN },
-	{    "cz", "cs_CZ", "Czech", CZ_CZE },
-	{    "nl", "nl_NL", "Dutch", NL_NLD },
-	{    "en",    "en", "English", EN_ANY }, // Generic English (when only one game version exist)
-	{    "gb", "en_GB", "English (GB)", EN_GRB },
-	{    "us", "en_US", "English (US)", EN_USA },
-	{    "fr", "fr_FR", "French", FR_FRA },
-	{    "de", "de_DE", "German", DE_DEU },
-	{    "gr", "el_GR", "Greek", GR_GRE },
-	{    "he", "he_IL", "Hebrew", HE_ISR },
-	{    "hb", "he_IL", "Hebrew", HE_ISR }, // Deprecated
-	{    "hr", "hr_HR", "Croatian", HR_HRV },
-	{    "hu", "hu_HU", "Hungarian", HU_HUN },
-	{    "it", "it_IT", "Italian", IT_ITA },
-	{    "jp", "ja_JP", "Japanese", JA_JPN },
-	{    "kr", "ko_KR", "Korean", KO_KOR },
-	{    "nb", "nb_NO", "Norwegian Bokm\xE5l", NB_NOR }, // TODO Someone should verify the unix locale
-	{    "pl", "pl_PL", "Polish", PL_POL },
-	{    "br", "pt_BR", "Portuguese", PT_BRA },
-	{    "ru", "ru_RU", "Russian", RU_RUS },
-	{    "es", "es_ES", "Spanish", ES_ESP },
-	{    "se", "sv_SE", "Swedish", SE_SWE },
-	{       0,       0, 0, UNK_LANG }
-};
-
-Language parseLanguage(const String &str) {
-	if (str.empty())
-		return UNK_LANG;
-
-	const LanguageDescription *l = g_languages;
-	for (; l->code; ++l) {
-		if (str.equalsIgnoreCase(l->code))
-			return l->id;
-	}
-
-	return UNK_LANG;
-}
-
-Language parseLanguageFromLocale(const char *locale) {
-	if (!locale || !*locale)
-		return UNK_LANG;
-
-	const LanguageDescription *l = g_languages;
-	for (; l->code; ++l) {
-		if (!strcmp(l->unixLocale, locale))
-			return l->id;
-	}
-
-	return UNK_LANG;
-}
-
-const char *getLanguageCode(Language id) {
-	const LanguageDescription *l = g_languages;
-	for (; l->code; ++l) {
-		if (l->id == id)
-			return l->code;
-	}
-	return 0;
-}
-
-const char *getLanguageLocale(Language id) {
-	const LanguageDescription *l = g_languages;
-	for (; l->code; ++l) {
-		if (l->id == id)
-			return l->unixLocale;
-	}
-	return 0;
-}
-
-const char *getLanguageDescription(Language id) {
-	const LanguageDescription *l = g_languages;
-	for (; l->code; ++l) {
-		if (l->id == id)
-			return l->description;
-	}
-	return 0;
-}
-
-
-#pragma mark -
-
-
-const PlatformDescription g_platforms[] = {
-	{ "2gs", "2gs", "2gs", "Apple IIgs", kPlatformApple2GS },
-	{ "3do", "3do", "3do", "3DO", kPlatform3DO },
-	{ "acorn", "acorn", "acorn", "Acorn", kPlatformAcorn },
-	{ "amiga", "ami", "amiga", "Amiga", kPlatformAmiga },
-	{ "atari", "atari-st", "st", "Atari ST", kPlatformAtariST },
-	{ "c64", "c64", "c64", "Commodore 64", kPlatformC64 },
-	{ "pc", "dos", "ibm", "DOS", kPlatformPC },
-	{ "pc98", "pc98", "pc98", "PC-98", kPlatformPC98 },
-	{ "wii", "wii", "wii", "Nintendo Wii", kPlatformWii },
-	{ "coco3", "coco3", "coco3", "CoCo3", kPlatformCoCo3 },
-
-	// The 'official' spelling seems to be "FM-TOWNS" (e.g. in the Indy4 demo).
-	// However, on the net many variations can be seen, like "FMTOWNS",
-	// "FM TOWNS", "FmTowns", etc.
-	{ "fmtowns", "towns", "fm", "FM-TOWNS", kPlatformFMTowns },
-
-	{ "linux", "linux", "linux", "Linux", kPlatformLinux },
-	{ "macintosh", "mac", "mac", "Macintosh", kPlatformMacintosh },
-	{ "pce", "pce", "pce", "PC-Engine", kPlatformPCEngine },
-	{ "nes", "nes", "nes", "NES", kPlatformNES },
-	{ "segacd", "segacd", "sega", "SegaCD", kPlatformSegaCD },
-	{ "windows", "win", "win", "Windows", kPlatformWindows },
-	{ "playstation", "psx", "psx", "Sony PlayStation", kPlatformPSX },
-	{ "cdi", "cdi", "cdi", "Philips CD-i", kPlatformCDi },
-	{ "ios", "ios", "ios", "Apple iOS", kPlatformIOS },
-
-	{ 0, 0, 0, "Default", kPlatformUnknown }
-};
-
-Platform parsePlatform(const String &str) {
-	if (str.empty())
-		return kPlatformUnknown;
-
-	// Handle some special case separately, for compatibility with old config
-	// files.
-	if (str == "1")
-		return kPlatformAmiga;
-	else if (str == "2")
-		return kPlatformAtariST;
-	else if (str == "3")
-		return kPlatformMacintosh;
-
-	const PlatformDescription *l = g_platforms;
-	for (; l->code; ++l) {
-		if (str.equalsIgnoreCase(l->code) || str.equalsIgnoreCase(l->code2) || str.equalsIgnoreCase(l->abbrev))
-			return l->id;
-	}
-
-	return kPlatformUnknown;
-}
-
-
-const char *getPlatformCode(Platform id) {
-	const PlatformDescription *l = g_platforms;
-	for (; l->code; ++l) {
-		if (l->id == id)
-			return l->code;
-	}
-	return 0;
-}
-
-const char *getPlatformAbbrev(Platform id) {
-	const PlatformDescription *l = g_platforms;
-	for (; l->code; ++l) {
-		if (l->id == id)
-			return l->abbrev;
-	}
-	return 0;
-}
-
-const char *getPlatformDescription(Platform id) {
-	const PlatformDescription *l = g_platforms;
-	for (; l->code; ++l) {
-		if (l->id == id)
-			return l->description;
-	}
-	return l->description;
-}
-
-
-#pragma mark -
-
-
 #define ENSURE_ASCII_CHAR(c) \
 		if (c < 0 || c > 127) \
 			return false
