+++ common/util.h
@@ -165,102 +165,6 @@ bool isSpace(int c);
  */
 bool isUpper(int c);
 
-
-/**
- * List of game language.
- */
-enum Language {
-	ZH_CNA,
-	ZH_TWN,
-	CZ_CZE,
-	NL_NLD,
-	EN_ANY,     // Generic English (when only one game version exist)
-	EN_GRB,
-	EN_USA,
-	FR_FRA,
-	DE_DEU,
-	GR_GRE,
-	HE_ISR,
-	HR_HRV,
-	HU_HUN,
-	IT_ITA,
-	JA_JPN,
-	KO_KOR,
-	NB_NOR,
-	PL_POL,
-	PT_BRA,
-	RU_RUS,
-	ES_ESP,
-	SE_SWE,
-
-	UNK_LANG = -1	// Use default language (i.e. none specified)
-};
-
-struct LanguageDescription {
-	const char *code;
-	const char *unixLocale;
-	const char *description;
-	Language id;
-};
-
-extern const LanguageDescription g_languages[];
-
-
-/** Convert a string containing a language name into a Language enum value. */
-extern Language parseLanguage(const String &str);
-extern Language parseLanguageFromLocale(const char *locale);
-extern const char *getLanguageCode(Language id);
-extern const char *getLanguageLocale(Language id);
-extern const char *getLanguageDescription(Language id);
-
-/**
- * List of game platforms. Specifying a platform for a target can be used to
- * give the game engines a hint for which platform the game data file are.
- * This may be optional or required, depending on the game engine and the
- * game in question.
- */
-enum Platform {
-	kPlatformPC,
-	kPlatformAmiga,
-	kPlatformAtariST,
-	kPlatformMacintosh,
-	kPlatformFMTowns,
-	kPlatformWindows,
-	kPlatformNES,
-	kPlatformC64,
-	kPlatformCoCo3,
-	kPlatformLinux,
-	kPlatformAcorn,
-	kPlatformSegaCD,
-	kPlatform3DO,
-	kPlatformPCEngine,
-	kPlatformApple2GS,
-	kPlatformPC98,
-	kPlatformWii,
-	kPlatformPSX,
-	kPlatformCDi,
-	kPlatformIOS,
-
-	kPlatformUnknown = -1
-};
-
-struct PlatformDescription {
-	const char *code;
-	const char *code2;
-	const char *abbrev;
-	const char *description;
-	Platform id;
-};
-
-extern const PlatformDescription g_platforms[];
-
-/** Convert a string containing a platform name into a Platform enum value. */
-extern Platform parsePlatform(const String &str);
-extern const char *getPlatformCode(Platform id);
-extern const char *getPlatformAbbrev(Platform id);
-extern const char *getPlatformDescription(Platform id);
-
 }	// End of namespace Common
-
 
 #endif
