+++ engines/engine.h
@@ -24,6 +24,8 @@
 
 #include "common/scummsys.h"
 #include "common/str.h"
+#include "common/language.h"
+#include "common/platform.h"
 
 class OSystem;
