+++ graphics/sjis.h
@@ -43,7 +43,7 @@
 #endif
 
 #include "common/scummsys.h"
-#include "common/util.h"
+#include "common/platform.h"
 
 namespace Graphics {
